+++ staffer-react-mobx/src/components/Button/Button.css
@@ -1,4 +1,4 @@
-@import 'suitcss-components-button';
+@import "suitcss-components-button";
 @import "styles/theme.css";
 @import "styles/theme-map.css";
