+++ staffer-react-mobx/src/components/Button/package.json
@@ -1,6 +1,6 @@
 {
   "name": "Button",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "main": "./Button.js"
 }
