+++ staffer-react-mobx/src/components/H1/package.json
@@ -1,6 +1,6 @@
 {
   "name": "H1",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "main": "./H1.js"
 }
