+++ staffer-react-mobx/src/components/Header/Header.css
@@ -1,15 +1,16 @@
 /** @define Header */
 
 @import "styles/theme.css";
+@import "styles/theme-map.css";
+
 
 .Header {
-
     padding: 1rem .5rem;
-    color: var(--color-white);
-    background: var(--color-twitter-blue);
-    position: fixed;
+    color: var(--Header-color);
+    background: var(--Header-background-color);
+    position: var(--Header-position);
     width: 100%;
-    height: 4rem;
+    height: var(--Header-height);
     z-index: 100;
 }
