+++ staffer-react-mobx/src/components/Input/Input.css
@@ -3,9 +3,4 @@
 
 /** @define Input */
 
-.Input {
-    /*font-weight: var(--H1-font-weight);*/
-    /*font-family: var(--H1-font-family);*/
-    /*font-size: var(--H1-font-size);*/
-    /*margin-bottom: var(--H1-margin-bottom);*/
-}
+.Input {}
