+++ staffer-react-mobx/src/components/Input/Input.js
@@ -5,7 +5,7 @@
  */
 
 
-import React from 'react';
+import React, {PropTypes} from 'react';
 import './Input.css';
 
 
@@ -19,5 +19,15 @@ const Input = ({id, label, placeholder, type, defaultValue, handleAction, ariaLa
                aria-label={ariaLabel} />
     </div>
 );
+
+Input.propTypes = {
+    id: PropTypes.string,
+    label: PropTypes.string,
+    placeholder: PropTypes.string,
+    type: PropTypes.string,
+    defaultValue: PropTypes.string,
+    ariaLabel: PropTypes.string,
+    handleAction: PropTypes.func
+};
 
 export default Input;
