+++ staffer-react-mobx/src/components/Input/package.json
@@ -1,6 +1,6 @@
 {
   "name": "Input",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "main": "./Input.js"
 }
