+++ staffer-react-mobx/src/components/Main/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Layout",
+  "name": "Main",
   "version": "0.0.0",
   "private": true,
   "main": "./Main.js"
