+++ staffer-react-mobx/src/components/NeedsPanel/package.json
@@ -1,6 +1,6 @@
 {
   "name": "NeedsPanel",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "main": "./NeedsPanel.js"
 }
