+++ staffer-react-mobx/src/components/PeoplePanel/PeopleTable.js
