+++ staffer-react-mobx/src/components/PeoplePanel/package.json
