+++ staffer-react-mobx/src/components/Select/package.json
@@ -1,6 +1,6 @@
 {
   "name": "Select",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "main": "./Select.js"
 }
