--- staffer-react-mobx/src/containers/NeedsFilter/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "NeedsFilter",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./NeedsFilter.js"
-}
