--- staffer-react-mobx/src/containers/NeedsTable/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "NeedsTable",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./NeedsTable.js"
-}
