--- staffer-react-mobx/src/containers/PeopleTable/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "PeopleTeble",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./PeopleTable.js"
-}
