+++ staffer-react-mobx/src/main.js
@@ -3,13 +3,13 @@ import './styles/base.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Layout from './containers/Main';
+import Main from './components/Main';
 import ContextProvider from './ContextProvider';
 import store from './store';
 
 let root = (
     <ContextProvider context={{store}}>
-        <Layout />
+        <Main />
     </ContextProvider>
 );
