+++ staffer-react-mobx/src/styles/base.css
@@ -50,6 +50,13 @@ html {
     letter-spacing: 0.05em;
 }
 
+html, body {
+    width: 100%;
+    height: 100%;
+    margin: 0;
+}
+
+
 h1,
 h2,
 h3 {
